Reorganize demonstration as examples with dev dependencies
