+++ Cargo.toml
@@ -16,15 +16,15 @@ license = "MIT"
 members = ["columnar_derive"]
 
 [dependencies]
-rmp-serde = "1.3.0"
-bincode = "1.3.3"
 serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
 bytemuck = "1.20"
 columnar_derive = { path = "columnar_derive", version = "0.2" }
 
 [dev-dependencies]
 bencher = "0.1.5"
+bincode = "1.3.3"
+rmp-serde = "1.3.0"
+serde_json = "1.0"
 
 [profile.test]
 opt-level = 2
