+++ src/adts/mod.rs
@@ -3,5 +3,4 @@
 //! These ADTs are generally recursive, which is why they required a special implementation.
 
 pub mod tree;
-pub mod json;
 pub mod art;
